마이페이지에 점수 추가 완료
